upgrade

+++ backend/service/k8s/cronjob.go
@@ -2,12 +2,15 @@ package k8s
 
 import (
 	"context"
+	"fmt"
+	"strconv"
 	"strings"
 
 	"github.com/golang/protobuf/ptypes/wrappers"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
-	v1beta1 "k8s.io/api/batch/v1beta1"
+	"k8s.io/api/batch/v1"
+	"k8s.io/api/batch/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	k8sapiv1 "github.com/lyft/clutch/backend/api/k8s/v1"
@@ -19,16 +22,39 @@ func (s *svc) DescribeCronJob(ctx context.Context, clientset, cluster, namespace
 		return nil, err
 	}
 
-	cronJobs, err := cs.BatchV1beta1().CronJobs(cs.Namespace()).List(ctx, metav1.ListOptions{
+	major, minor, err := fetchVersion(cs)
+	if err != nil {
+		return nil, err
+	}
+
+	if major == 1 && minor < 21 {
+		cronJobs, err := cs.BatchV1beta1().CronJobs(cs.Namespace()).List(ctx, metav1.ListOptions{
+			FieldSelector: "metadata.name=" + name,
+		})
+		if err != nil {
+			return nil, err
+		}
+
+		if len(cronJobs.Items) == 1 {
+			return ProtoForV1Beta1CronJob(cs.Cluster(), &cronJobs.Items[0]), nil
+		}
+		if len(cronJobs.Items) > 1 {
+			return nil, status.Error(codes.FailedPrecondition, "located multiple cron jobs")
+		}
+		return nil, status.Error(codes.NotFound, "unable to locate specified cron job")
+
+	}
+
+	cronJobs, err := cs.BatchV1().CronJobs(cs.Namespace()).List(ctx, metav1.ListOptions{
 		FieldSelector: "metadata.name=" + name,
 	})
 	if err != nil {
 		return nil, err
 	}
-
 	if len(cronJobs.Items) == 1 {
 		return ProtoForCronJob(cs.Cluster(), &cronJobs.Items[0]), nil
-	} else if len(cronJobs.Items) > 1 {
+	}
+	if len(cronJobs.Items) > 1 {
 		return nil, status.Error(codes.FailedPrecondition, "located multiple cron jobs")
 	}
 	return nil, status.Error(codes.NotFound, "unable to locate specified cron job")
@@ -45,7 +71,25 @@ func (s *svc) ListCronJobs(ctx context.Context, clientset, cluster, namespace st
 		return nil, err
 	}
 
-	cronJobList, err := cs.BatchV1beta1().CronJobs(cs.Namespace()).List(ctx, opts)
+	major, minor, err := fetchVersion(cs)
+	if err != nil {
+		return nil, err
+	}
+
+	if major == 1 && minor < 21 {
+		cronJobList, err := cs.BatchV1beta1().CronJobs(cs.Namespace()).List(ctx, opts)
+		if err != nil {
+			return nil, err
+		}
+		var cronJobs []*k8sapiv1.CronJob
+		for _, d := range cronJobList.Items {
+			cronJob := d
+			cronJobs = append(cronJobs, ProtoForV1Beta1CronJob(cs.Cluster(), &cronJob))
+		}
+		return cronJobs, nil
+	}
+
+	cronJobList, err := cs.BatchV1().CronJobs(cs.Namespace()).List(ctx, opts)
 	if err != nil {
 		return nil, err
 	}
@@ -55,7 +99,6 @@ func (s *svc) ListCronJobs(ctx context.Context, clientset, cluster, namespace st
 		cronJob := d
 		cronJobs = append(cronJobs, ProtoForCronJob(cs.Cluster(), &cronJob))
 	}
-
 	return cronJobs, nil
 }
 
@@ -66,10 +109,19 @@ func (s *svc) DeleteCronJob(ctx context.Context, clientset, cluster, namespace, 
 	}
 
 	opts := metav1.DeleteOptions{}
-	return cs.BatchV1beta1().CronJobs(cs.Namespace()).Delete(ctx, name, opts)
+
+	major, minor, err := fetchVersion(cs)
+	if err != nil {
+		return err
+	}
+
+	if major == 1 && minor < 21 {
+		return cs.BatchV1beta1().CronJobs(cs.Namespace()).Delete(ctx, name, opts)
+	}
+	return cs.BatchV1().CronJobs(cs.Namespace()).Delete(ctx, name, opts)
 }
 
-func ProtoForCronJob(cluster string, k8scronJob *v1beta1.CronJob) *k8sapiv1.CronJob {
+func ProtoForV1Beta1CronJob(cluster string, k8scronJob *v1beta1.CronJob) *k8sapiv1.CronJob {
 	clusterName := GetKubeClusterName(k8scronJob)
 	if clusterName == "" {
 		clusterName = cluster
@@ -99,4 +151,54 @@ func ProtoForCronJob(cluster string, k8scronJob *v1beta1.CronJob) *k8sapiv1.Cron
 		ret.StartingDeadlineSeconds = &wrappers.Int64Value{Value: *k8scronJob.Spec.StartingDeadlineSeconds}
 	}
 	return ret
+}
+
+func ProtoForCronJob(cluster string, k8scronJob *v1.CronJob) *k8sapiv1.CronJob {
+	clusterName := GetKubeClusterName(k8scronJob)
+	if clusterName == "" {
+		clusterName = cluster
+	}
+	// Required fields
+	ret := &k8sapiv1.CronJob{
+		Cluster:     clusterName,
+		Namespace:   k8scronJob.Namespace,
+		Name:        k8scronJob.Name,
+		Schedule:    k8scronJob.Spec.Schedule,
+		Labels:      k8scronJob.Labels,
+		Annotations: k8scronJob.Annotations,
+	}
+
+	// Update optional fields
+	if k8scronJob.Spec.Suspend != nil {
+		ret.Suspend = *k8scronJob.Spec.Suspend
+	}
+	if k8scronJob.Spec.ConcurrencyPolicy != "" {
+		ret.ConcurrencyPolicy = k8sapiv1.CronJob_ConcurrencyPolicy(
+			k8sapiv1.CronJob_ConcurrencyPolicy_value[strings.ToUpper(string(k8scronJob.Spec.ConcurrencyPolicy))])
+	}
+	if k8scronJob.Status.Active != nil {
+		ret.NumActiveJobs = int32(len(k8scronJob.Status.Active))
+	}
+	if k8scronJob.Spec.StartingDeadlineSeconds != nil {
+		ret.StartingDeadlineSeconds = &wrappers.Int64Value{Value: *k8scronJob.Spec.StartingDeadlineSeconds}
+	}
+	return ret
+}
+
+func fetchVersion(cs ContextClientset) (int, int, error) {
+	version, err := cs.Discovery().ServerVersion()
+	if err != nil {
+		return 0, 0, fmt.Errorf("unable to get kubernetes server version info: %w", err)
+	}
+
+	major, err := strconv.Atoi(version.Major)
+	if err != nil {
+		return 0, 0, fmt.Errorf("unable to convert kube major version to int: %w", err)
+	}
+
+	minor, err := strconv.Atoi(strings.Trim(version.Minor, "+"))
+	if err != nil {
+		return 0, 0, fmt.Errorf("unable to convert kube minor version to int: %w", err)
+	}
+	return major, minor, nil
 }
